+++ client/src/Compornent/Card_Gwey.jsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import Card from '@material-ui/core/Card';
+import CardActionArea from '@material-ui/core/CardActionArea';
+import CardMedia from '@material-ui/core/CardMedia';
+
+const useStyles = makeStyles({
+  root: {
+    maxWidth: 300,
+  },
+  media: {
+    width:300,
+    height: 300,
+  },
+});
+
+const Card_Gwey = (props) => {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.media}>
+       {props.contents}
+    </div>
+  );
+}
+
+export default Card_Gwey;
