+++ client/src/index.js
@@ -3,10 +3,12 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
+import S_001 from './View/S_001'
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    {/* <App /> */}
+    <S_001 />
   </React.StrictMode>,
   document.getElementById('root')
 );
