+++ client/src/variables/data.jsx
@@ -3,41 +3,21 @@ export const getData =
     {
       "contents_id":1,
       "contents_detail_id":1,
-      "contents_name":"",
-      "contents_statement":"個人情報とは何を指すか。下記の4つの中から選びなさい。",
-      "choice1":"氏名",
-      "choice2":"電話番号",
-      "choice3":"住所",
-      "choice4":"勤務先",
-      "answer":3
+      "contents_statement":"../Assets/img/page1.jpg",
     },
     {
       "contents_id":1,
       "contents_detail_id":2,
-      "contents_name":"",
-      "contents_statement":"個人情報とは何を指すか。下記の4つの中から選びなさい②．",
-      "choice1":"名前",
-      "choice2":"FAX番号",
-      "choice3":"郵便番号",
-      "choice4":"職場",
-      "answer":3
+      "contents_statement":"../Assets/img/page2.jpg",
     },
     {
       "contents_id":1,
       "contents_detail_id":3,
-      "contents_name":"",
-      "contents_statement":"./image/slide1.jpg",
-      "choice1":"名前",
-      "choice2":"FAX番号",
-      "choice3":"郵便番号",
-      "choice4":"職場",
-      "answer":3
-    },
-    {
-      "contents_detail_id":4,
-      "contents_statement":"./img/default.jpg"
+      "contents_statement":"../Assets/img/page3.jpg",
     }
-    // {
+  ];
+
+  // {
     //   "CustomerID":1,
     //   "CustomerCode":"3001",
     //   "CustomerName":"株式会社ノア",
@@ -61,4 +41,3 @@ export const getData =
     //   "CustomerClassID":2,
     //   "PrefecturalID":4
     // }
-  ];
